modules/5_defense_median/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
Pillow>=9.0.0
numpy>=1.21.0
